Rename CoursePopup
Old one to "Old", remove "2" from new one
