+++ src/views/components/injected/CoursePopup/CourseHeadingAndActions.tsx
