+++ src/views/components/injected/CoursePopupOld/CourseDescription/CourseDescription.module.scss
