+++ src/views/components/injected/CoursePopupOld/CourseDescription/CourseDescription.tsx
