+++ src/views/components/injected/CoursePopupOld/CourseHeader/CourseHeader.module.scss
