+++ src/views/components/injected/CoursePopupOld/CourseHeader/CourseHeader.tsx
