+++ src/views/components/injected/CoursePopupOld/CoursePopup.module.scss
