+++ src/views/components/injected/CoursePopupOld/GradeDistribution/GradeDistribution.module.scss
