+++ src/views/components/injected/CoursePopupOld/GradeDistribution/GradeDistribution.tsx
